Port the kernel to 64-bit long mode
Everything is x86-64 now: the kernel, the user programs, the interrupt
frames, the page tables and the SMP trampoline.

boot/boot.S gained the awkward part. Long mode cannot be entered from
32-bit protected mode directly, because the processor requires paging
to be on before the switch and paging in long mode needs four levels of
tables that have to exist first. So it builds tables identity mapping
the first 4 GiB with 2 MiB pages, turns on PAE, asks for long mode,
turns on paging, and jumps through a 64-bit descriptor. It also refuses
to continue on a processor without long mode rather than faulting
somewhere less explicable a few instructions later.

paging.c walks four levels instead of two. That change had a
consequence worth naming: one top level entry covers 512 GiB, and the
kernel's mappings are shared into every address space by copying that
entry, so a program at 1 GiB would have shared the kernel's entry and
put its own page tables inside the kernel's. Every address space would
have seen every other one's memory and none of it could be freed.
User space now gets an entry of its own at 512 GiB.

The self test caught that: "its address space was reclaimed" failed
because nothing was being reclaimed.

Two more truncations that the compiler could not warn about, because
both were explicit casts:

  - sys_win_surface returned an i32. It is the one system call whose
    result is a pointer, and half of it was being thrown away, so the
    program wrote to the low half and took a page fault. The whole
    syscall table returns a machine word now.
  - wintest printed eight hex digits, which made the truncated pointer
    look like a correct one. It prints all sixteen.

The AHCI and e1000 drivers set the upper half of every DMA address
rather than assuming zero, and the RTL8139 refuses rather than
truncating, because its buffer register genuinely has no upper half.

The SMP trampoline starts in real mode, where paging cannot be turned
on without protected mode already being on, so both bits go on in one
write. Doing them separately started none of the processors.

QEMU's -kernel will not take a 64-bit ELF, because multiboot predates
long mode. The multiboot header carries the a.out kludge now, which
tells a loader to stop reading ELF headers and copy the flat image
instead, so the fast test loop survives. build.sh emits both.

183 self test checks with -smp 4, 174 with one processor, and the
shell test passes.

+++ include/acpi.h
@@ -12,7 +12,7 @@
 
 typedef struct {
     bool found;                 /* the tables were there and made sense */
-    u32  lapic_base;            /* MMIO address of the local APIC */
+    u64  lapic_base;            /* MMIO address of the local APIC */
     u32  ncpus;
     u8   apic_id[ACPI_MAX_CPUS];
     u8   usable[ACPI_MAX_CPUS]; /* the firmware says this one can be started */
